requirements.txt
# ============================================
# 🚀 RL STOCK MARKET TRADING SYSTEM
# Dependencias Completas - Sistema Avanzado de Trading con IA
# ============================================

# ✅ CORE TRADING & DATA
# ------------------------------------
MetaTrader5==5.0.5120                    # Conexión MetaTrader (opcional)
python-dotenv==1.0.0                     # Variables de entorno
yfinance==0.2.40                         # Datos Yahoo Finance (SP500)
ccxt==4.4.19                             # Binance API para Bitcoin

# ✅ DATA PROCESSING & ANALYSIS
# ------------------------------------
pandas==2.2.3                            # Análisis de datos
numpy==1.26.4                            # Operaciones numéricas
scipy==1.11.4                            # Cálculos científicos

# ✅ TECHNICAL ANALYSIS & INDICATORS
# ------------------------------------
ta==0.11.0                               # Indicadores técnicos básicos
pandas-ta==0.3.14b0                     # Indicadores técnicos avanzados
TA-Lib==0.4.32                          # Biblioteca técnica profesional

# ✅ MACHINE LEARNING & REINFORCEMENT LEARNING
# ------------------------------------
torch>=2.1.2                            # PyTorch para RL
stable-baselines3==2.6.0                # Algoritmos RL (DQN, A2C, PPO, SAC)
gymnasium==1.1.1                        # Entornos RL modernos
gym==0.26.2                             # Compatibilidad gym legacy
tensorboard==2.15.1                     # Visualización de entrenamiento
optuna==3.5.0                           # Optimización de hiperparámetros
scikit-learn==1.3.2                     # ML tradicional y métricas
tensorflow>=2.15.0                      # TensorFlow para modelos avanzados
keras>=3.0.0                            # API de alto nivel

# ✅ VISUALIZATIONS & DASHBOARD
# ------------------------------------
matplotlib==3.10.1                      # Gráficos base
seaborn==0.13.0                         # Gráficos estadísticos
plotly==5.18.0                          # Gráficos interactivos
dash==2.18.1                            # Dashboard web interactivo
dash-bootstrap-components==1.6.0        # Componentes UI para Dash

# ✅ JUPYTER & INTERACTIVE ANALYSIS
# ------------------------------------
jupyter==1.0.0                          # Notebook Jupyter
ipywidgets==8.1.5                       # Widgets interactivos
notebook==7.2.2                         # Interfaz notebook
jupyterlab==4.2.5                       # JupyterLab IDE

# ✅ DATABASE & STORAGE
# ------------------------------------
psycopg2-binary==2.9.9                  # PostgreSQL
sqlalchemy==2.0.23                      # ORM
alembic==1.13.1                         # Migraciones DB
sqlite3                                 # SQLite (incluido en Python)

# ✅ API & WEB SERVICES
# ------------------------------------
fastapi==0.108.0                        # API REST
uvicorn==0.25.0                         # Servidor ASGI
requests==2.31.0                        # HTTP requests
aiohttp==3.9.1                          # Async HTTP
websocket-client==1.7.0                 # WebSocket client

# ✅ ASYNC & STREAMING
# ------------------------------------
asyncio                                 # Async (incluido en Python)
confluent-kafka==2.3.0                  # Kafka streaming
redis==5.0.8                            # Cache y mensaje

# ✅ MONITORING & ALERTAS
# ------------------------------------
prometheus-client==0.19.0               # Métricas Prometheus
grafana-api==1.0.3                      # Dashboards Grafana
slack-sdk==3.33.1                       # Notificaciones Slack
discord.py==2.4.0                       # Notificaciones Discord

# ✅ MATHEMATICAL & NUMERICAL
# ------------------------------------
numba==0.60.0                           # Compilación JIT para velocidad
cython==3.0.11                          # Extensiones C
sympy==1.13.3                           # Matemáticas simbólicas

# ✅ CONFIGURATION & UTILITIES
# ------------------------------------
pyyaml==6.0.1                           # Archivos YAML
click==8.1.7                            # CLI interfaces
schedule==1.2.0                         # Tareas programadas
python-json-logger==2.0.7               # Logging JSON
rich==13.9.2                            # Terminal colorido
tqdm==4.66.6                            # Barras de progreso

# ✅ DEVELOPMENT & TESTING
# ------------------------------------
pytest==7.4.3                           # Testing framework
pytest-cov==6.0.0                       # Cobertura de tests
black==23.12.1                          # Formateo de código
flake8==7.0.0                           # Linting
mypy==1.13.0                            # Type checking
pre-commit==4.0.1                       # Git hooks

# ✅ SECURITY & ENCRYPTION
# ------------------------------------
cryptography==43.0.3                    # Criptografía
pycryptodome==3.21.0                    # Cifrado avanzado
keyring==25.4.1                         # Gestión de claves

# ✅ DEPLOYMENT & CONTAINERIZATION
# ------------------------------------
docker==7.1.0                           # Docker Python API
kubernetes==31.0.0                      # Kubernetes API
gunicorn==23.0.0                        # WSGI server

# ✅ FINANCIAL & TRADING SPECIFIC
# ------------------------------------
quantlib==1.36                          # Biblioteca financiera cuantitativa
zipline-reloaded==3.0.4                 # Backtesting framework
pyfolio==0.9.2                          # Análisis de performance
empyrical==0.5.5                        # Métricas financieras
bt==1.0.1                               # Backtesting framework ligero

# ✅ NOTIFICATIONS & COMMUNICATIONS
# ------------------------------------
twilio==9.3.6                           # SMS/llamadas
sendgrid==6.11.0                        # Email
telegram-bot-api==7.9.1                 # Telegram bot

# ============================================
# 📋 NOTAS DE INSTALACIÓN:
# ============================================
# 
# 🔧 Instalación básica:
# pip install -r requirements.txt
# 
# 🚀 Para desarrollo completo:
# python utils/install_advanced_dependencies.py
# 
# ⚠️ Dependencias problemáticas en Windows:
# - TA-Lib: Requiere Visual Studio Build Tools
# - numba: Puede fallar en algunos sistemas
# - MetaTrader5: Solo funciona en Windows
# 
# 🐧 Para Linux/Mac:
# sudo apt-get install build-essential (Ubuntu)
# brew install gcc (macOS)
# 
# 📊 Para uso mínimo (solo análisis):
# pip install pandas numpy matplotlib yfinance scikit-learn
# 
# 🤖 Para ML completo:
# pip install torch stable-baselines3 gymnasium tensorflow
# 
# 📈 Para trading en vivo:
# pip install ccxt python-dotenv MetaTrader5
# 
# ============================================
